Fix last digit popping in increment_integer()
There was an error in the way the last digit is popped:

    "#{h}#{String.slice(digs, 0..(String.length(digs) - 1))}"

The 0..(length - 1) range actually represents the whole string.

Popping the last digit is better expressed with String.slice/3:

    "#{h}#{String.slice(digs, 0, String.length(digs) - 1)}"

This fixes an issue where generating an index between "Yzz" and "Z0G"
was producing the index "Z00", which is an invalid fractional index. The
code now produces "Z0", which is correct.